+++ apps/restconf/restconf_root.c
@@ -432,6 +432,7 @@ api_root_restconf(clicon_handle        h,
     cbuf          *cb = NULL;
     char          *media_str = NULL;
     restconf_media media_out = YANG_DATA_JSON;
+    restconf_media media_in;
     char          *indata = NULL;
     char          *username = NULL;
     int            ret;
@@ -468,7 +469,6 @@ api_root_restconf(clicon_handle        h,
 	    goto ok;
 	}
     }
-
     clicon_debug(1, "%s ACCEPT: %s %s", __FUNCTION__, media_str, restconf_media_int2str(media_out));
 
     if ((pvec = clicon_strsep(path, "/", &pn)) == NULL)