+++ lib/src/clixon_netconf_lib.c
@@ -1526,6 +1526,10 @@ netconf_module_load(clicon_handle h)
     if (clicon_option_bool(h, "CLICON_NETCONF_MESSAGE_ID_OPTIONAL") == 1)
 	xml_bind_netconf_message_id_optional(1);
 #endif
+    /* Load restconf collection */
+    if (yang_spec_parse_module(h, "ietf-netconf-collection", NULL, yspec)< 0)
+	goto done;    
+
     retval = 0;
  done:
     return retval;